+++ controller/src/main/java/com/linbit/linstor/api/rest/v1/ResourceGroups.java
@@ -358,7 +358,7 @@ public class ResourceGroups
             .flatMap(apiCallRcWith ->
             {
                 Response resp;
-                JsonGenTypes.QuerySizeInfoResponse qsiResp = Json.pojoToQuersSizeInfoResp(
+                JsonGenTypes.QuerySizeInfoResponse qsiResp = Json.pojoToQuerySizeInfoResp(
                     apiCallRcWith.getValue(),
                     apiCallRcWith.getApiCallRc()
                 );