+++ server/src/main/java/com/linbit/linstor/api/pojo/QueryAllSizeInfoRequestPojo.java
@@ -1,0 +1,16 @@
+package com.linbit.linstor.api.pojo;
+
+public class QueryAllSizeInfoRequestPojo
+{
+    private final AutoSelectFilterPojo autoSelectFilterData;
+
+    public QueryAllSizeInfoRequestPojo(AutoSelectFilterPojo autoSelectFilterDataRef)
+    {
+        autoSelectFilterData = autoSelectFilterDataRef;
+    }
+
+    public AutoSelectFilterPojo getAutoSelectFilterData()
+    {
+        return autoSelectFilterData;
+    }
+}